+++ src/Raven.CodeAnalysis/Binder/BinderFactory.cs
@@ -31,6 +31,7 @@ class BinderFactory
             ArrowExpressionClauseSyntax => CreateBlockBinder(parentBinder),
             IfExpressionSyntax expr => new LocalScopeBinder(parentBinder!),
             IfStatementSyntax stmt => new LocalScopeBinder(parentBinder!),
+            IfPatternStatementSyntax stmt => new LocalScopeBinder(parentBinder!),
             ElseClauseSyntax elseClause => new LocalScopeBinder(parentBinder!),
             WhileStatementSyntax stmt => new BlockBinder(parentBinder!.ContainingSymbol, parentBinder!),
             ForStatementSyntax stmt => new BlockBinder(parentBinder!.ContainingSymbol, parentBinder!),